requirements.txt
Authlib
celery
Django
djangorestframework
drf-yasg
kafka-python
python-logstash
psycopg2-binary
python-memcached
redis
requests
uWSGI
